face_recognition/pythonProject/requirements.txt
opencv-python~=4.9.0.80
numpy~=1.26.4
pillow~=10.2.0
torchvision~=0.17.0